requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
MouseInfo==0.1.3
pillow==10.4.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
sqlparse==0.5.1
tzdata==2024.1
